+++ src/main/java/org/truffleruby/core/format/format/FormatFFloatNode.java
@@ -18,8 +18,6 @@ package org.truffleruby.core.format.format;
 
 import java.math.BigInteger;
 import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-import java.util.Locale;
 
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.dsl.ImportStatic;
@@ -28,6 +26,7 @@ import com.oracle.truffle.api.dsl.Specialization;
 import org.truffleruby.collections.ByteArrayBuilder;
 import org.truffleruby.core.format.printf.PrintfSimpleTreeBuilder;
 import org.truffleruby.core.string.StringOperations;
+import org.truffleruby.core.string.StringUtils;
 
 @ImportStatic(Double.class)
 public abstract class FormatFFloatNode extends FormatFloatGenericNode {
@@ -56,8 +55,7 @@ public abstract class FormatFFloatNode extends FormatFloatGenericNode {
         final byte[] digits;
         DecimalFormat format = getLanguage().getCurrentThread().formatFFloat;
         if (format == null) {
-            final DecimalFormatSymbols formatSymbols = new DecimalFormatSymbols(Locale.ENGLISH);
-            format = new DecimalFormat("", formatSymbols);
+            format = StringUtils.newDecimalFormat("");
             getLanguage().getCurrentThread().formatFFloat = format;
         }