+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Compatibility:
 * Promote `File#path` and `File#to_path` to `IO#path` and `IO#to_path` and make IO#new accept an optional `path:` keyword argument (#3039, @moste00)
 * Display "unhandled exception" as the message for `RuntimeError` instances with an empty message (#3255, @nirvdrum).
 * Set `RbConfig::CONFIG['configure_args']` for openssl and libyaml (#3170, #3303, @eregon).
+* Use "half away from zero" rounding for `Kernel#{sprintf,format}` like CRuby (#3360, @eregon).
 
 Performance: